heroku fix 5

+++ models/models.go
@@ -7,7 +7,7 @@ type Country struct {
 	ID          primitive.ObjectID `bson:"_id,omitempty"`
 	Country     string             `bson:"country,omitempty"`
 	CountryInfo struct {
-		ID   int32   `bson:"_id,omitempty"`
+		ID   int64   `bson:"_id,omitempty"`
 		Iso2 string  `bson:"iso2,omitempty"`
 		Iso3 string  `bson:"iso3,omitempty"`
 		Lat  float64 `bson:"lat,omitempty"`
@@ -15,19 +15,19 @@ type Country struct {
 		Flag string  `bson:"flag,omitempty"`
 	}
 	Updated             int64   `bson:"updated,omitempty"`
-	Cases               int32   `bson:"cases,omitempty"`
-	TodayCases          int32   `bson:"todayCases,omitempty"`
-	Deaths              int32   `bson:"deaths,omitempty"`
-	TodayDeaths         int32   `bson:"todayDeaths,omitempty"`
-	Recovered           int32   `bson:"recovered,omitempty"`
-	Active              int32   `bson:"active,omitempty"`
-	Critical            int32   `bson:"critical,omitempty"`
+	Cases               int64   `bson:"cases,omitempty"`
+	TodayCases          int64   `bson:"todayCases,omitempty"`
+	Deaths              int64   `bson:"deaths,omitempty"`
+	TodayDeaths         int64   `bson:"todayDeaths,omitempty"`
+	Recovered           int64   `bson:"recovered,omitempty"`
+	Active              int64   `bson:"active,omitempty"`
+	Critical            int64   `bson:"critical,omitempty"`
 	CasesPerOneMillion  float64 `bson:"casesPerOneMillion,omitempty"`
 	DeathsPerOneMillion float64 `bson:"deathsPerOneMillion,omitempty"`
-	Tests               int32   `bson:"tests,omitempty,omitempty"`
+	Tests               int64   `bson:"tests,omitempty,omitempty"`
 	TestsPerOneMillion  float64 `bson:"testsPerOneMillion,omitempty"`
-	PopulationDensity   int32   `bson:"populationDensity"`
-	FortnightCases      []int32 `bson:"fortnightCases"`
+	PopulationDensity   int64   `bson:"populationDensity"`
+	FortnightCases      []int64 `bson:"fortnightCases"`
 	FortnightAverage    float64	`bson:"fortnightAverage"`
 	WeekAverage         float64 `bson:"weekAverage"`
 	DropRate            float64	`bson:"dropRate"`
